simx86: disassemble JIT-generated code for 'e' level>7

+++ src/base/lib/misc/dis8086.h
@@ -4,6 +4,6 @@
  * for details see file COPYING in the DOSEMU distribution
  */
 
-int  dis_8086(unsigned int, char *, int, unsigned int *, unsigned int);
+int  dis_8086(uintptr_t, char *, int, unsigned int *, unsigned int);
